refactor: 현재 구조에 맞게 info를 application layer로
